ORCH-0964: render image+GIF+video brand covers via shared EventCoverMedia [deploy]
Move EventCoverMedia + EventCover into @mingla/event-rendering so one component
backs mingla-business, app-mobile, and the shared brand page. Brand hero + cover
blocks now render image/GIF/video (web <video>/<img>, native expo-image/expo-video,
muted autoplay) instead of dropping video to a blank block. Business-app cover
files become re-export shims (cards unchanged). Add expo-video to app-mobile.
Repoint ORCH-0805 gate Check 9 to the shared brand page.

[TEST-MOD-APPROVED ORCH-0964] for the two locked test path updates.
