+++ app-mobile/app.json
@@ -128,6 +128,7 @@
       "expo-calendar",
       "expo-apple-authentication",
       "expo-av",
+      "expo-video",
       [
         "expo-build-properties",
         {
