+++ app-mobile/package.json
@@ -109,6 +109,7 @@
     "expo-location": "~19.0.8",
     "expo-router": "~6.0.23",
     "expo-sharing": "~14.0.7",
+    "expo-video": "~3.0.16",
     "expo-splash-screen": "~31.0.13",
     "expo-status-bar": "~3.0.9",
     "expo-symbols": "~1.0.8",
