+++ packages/brand-rendering/PublicBrandPage.tsx
@@ -21,6 +21,7 @@ import {
   Youtube,
 } from "lucide-react-native";
 import {
+  EventCoverMedia,
   MINGLA_DEFAULT_THEME,
   ThemeEntranceAnimation,
   resolveTheme,
@@ -36,6 +37,7 @@ import type {
   PublicBrandTicket,
   PublicBrandTrip,
   PublicBrandUpcoming,
+  PublicMediaType,
 } from "./types";
 
 type Tab = "upcoming" | "events" | "trips" | "experiences" | "about";
@@ -511,12 +513,13 @@ export const PublicBrandPage: React.FC<PublicBrandPageProps> = ({
         {brand.coverMediaUrl !== undefined &&
         brand.coverMediaUrl.length > 0 &&
         !coverMediaFailed ? (
-          <Image
-            source={{ uri: brand.coverMediaUrl }}
+          <EventCoverMedia
+            mediaUrl={brand.coverMediaUrl}
+            mediaType={brand.coverMediaType ?? null}
+            radius={0}
+            label="Brand cover"
             style={styles.heroGradient}
-            resizeMode="cover"
-            onError={() => setCoverMediaFailed(true)}
-            accessibilityLabel="Brand cover"
+            onMediaError={() => setCoverMediaFailed(true)}
           />
         ) : (
           <View style={[styles.heroGradient, { backgroundColor: heroColor }]} />
@@ -1361,26 +1364,16 @@ const CoverBlock: React.FC<{
   hue: number;
   mediaUrl: string | null;
   mediaType: string | null;
-}> = ({ hue, mediaUrl, mediaType }) => {
-  if (
-    mediaUrl !== null &&
-    mediaUrl.length > 0 &&
-    (mediaType === "image" || mediaType === "gif")
-  ) {
-    return (
-      <Image
-        source={{ uri: mediaUrl }}
-        style={styles.eventCover}
-        resizeMode="cover"
-      />
-    );
-  }
-  return (
-    <View
-      style={[styles.eventCover, { backgroundColor: `hsl(${hue}, 60%, 45%)` }]}
-    />
-  );
-};
+}> = ({ hue, mediaUrl, mediaType }) => (
+  <EventCoverMedia
+    hue={hue}
+    mediaUrl={mediaUrl}
+    mediaType={mediaType as PublicMediaType | null}
+    radius={0}
+    label="Cover"
+    style={styles.eventCover}
+  />
+);
 
 const AboutTab: React.FC<{
   brand: PublicBrand;
