+++ packages/event-rendering/index.ts
@@ -30,6 +30,16 @@ export type {
   QuantityRowTicket,
   QuantityRowTheme,
 } from "./QuantityRow";
+// ORCH-0964 — shared cover media (image + GIF + video, web + native) used by
+// mingla-business cards/public pages, app-mobile, AND the shared brand page.
+export { EventCover } from "./EventCover";
+export type { EventCoverProps } from "./EventCover";
+export { EventCoverMedia } from "./EventCoverMedia";
+export type {
+  EventCoverMediaProps,
+  EventCoverMediaErrorEvent,
+} from "./EventCoverMedia";
+export { resolveEventCoverMediaPresentation } from "./coverMediaPresentation";
 export type {
   PublicEventProps,
   PublicBrandProps,
