+++ packages/event-rendering/package.json
@@ -10,6 +10,10 @@
     "react": "*",
     "react-native": "*",
     "expo-blur": "*",
-    "lottie-react-native": "*"
+    "expo-image": "*",
+    "expo-linear-gradient": "*",
+    "expo-video": "*",
+    "lottie-react-native": "*",
+    "react-native-svg": "*"
   }
 }
